(feat:JS): export createQuery()
